backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
sqlalchemy==2.0.35
aiosqlite==0.20.0
pyyaml==6.0.2
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
python-pam==2.0.2
